Moved riot crud tests to serparate files
- Moved riot crud tests out of `crud.py` to their own respective files
- Updated crud tests to use crud instead of db_util
- Created new CRUD operation and tests for get player metadata and player stats
