+++ tests/db/test_crud_riot_league.py
